requirements.txt
torch==1.12.0
pytorch-lightning==1.6.4
einops==0.4.1
hydra-core==1.2.0
transformers==4.20.1